Modifiy python API architecture
- Merged messages and transports/serialization subpakages into flat modules.
- De-duplicated socket cleanup and nonce verfcation helpers acros the package.
