--- entity/python/iotauth/serialization/__init__.py
@@ -1,19 +1,0 @@
-"""Serialization helpers for IoTAuth binary protocol data."""
-
-from .binary import (
-    decode_uint_be,
-    decode_varint,
-    encode_uint_be,
-    encode_varint,
-    parse_frame,
-    serialize_frame,
-)
-
-__all__ = [
-    "decode_uint_be",
-    "decode_varint",
-    "encode_uint_be",
-    "encode_varint",
-    "parse_frame",
-    "serialize_frame",
-]
