--- entity/python/iotauth/transports/__init__.py
@@ -1,5 +1,0 @@
-"""Transport helpers for IoTAuth protocol frames."""
-
-from .tcp import connect, recv_frame, send_frame
-
-__all__ = ["connect", "recv_frame", "send_frame"]
